extract search and suggestions components from sidepanel
