+++ src/components/svelte/SidePanel.svelte
@@ -1,255 +1,100 @@
 <script lang="ts">
-  // import RoomDisplay from "./RoomDisplay.svelte";
   import { onMount } from "svelte";
-  import { debounce } from "es-toolkit/function";
-  import type { RoomData } from "../../lib/types";
-  import { modalStore } from "../../lib/store.svelte";
-  // currentRoomStore,
-  // import RoomModalContent from "./RoomModalContent.svelte";
-  import { getAppData } from "../../lib/context";
+  import Search from "./Search.svelte";
 
-  const { rooms, classesMap } = getAppData();
-
-  const MAX_DISPLAY_RESULT = 20;
-
-  let searchElement: HTMLInputElement | null = $state(null);
-  let searchInput: string = $state("");
-  let typing: boolean = $state(false);
-  let paginateOffset: number = $state(0);
+  // let searchElement: HTMLInputElement | null = $state(null);
+  // let searchInput: string = $state("");
+  // let paginateOffset: number = $state(0);
   // svelte-ignore state_referenced_locally
-  let roomsResult: typeof rooms = $state(rooms);
-  const paginatedRooms = $derived(
-    roomsResult.slice(
-      paginateOffset * MAX_DISPLAY_RESULT,
-      (paginateOffset + 1) * MAX_DISPLAY_RESULT,
-    ),
-  );
+  // let roomsResult: typeof rooms = $state(rooms);
+  // const paginatedRooms = $derived(
+  //   roomsResult.slice(
+  //     paginateOffset * MAX_DISPLAY_RESULT,
+  //     (paginateOffset + 1) * MAX_DISPLAY_RESULT,
+  //   ),
+  // );
 
-  const maxPaginateOffset: number = $derived(
-    Math.floor((roomsResult.length - 1) / MAX_DISPLAY_RESULT) + 1,
-  );
+  // const maxPaginateOffset: number = $derived(
+  //   Math.floor((roomsResult.length - 1) / MAX_DISPLAY_RESULT) + 1,
+  // );
 
   onMount(() => {
     const params = new URLSearchParams(window.location.search);
     const paramsQuery = params.get("s");
     if (paramsQuery != null) {
-      searchInput = paramsQuery;
-      const searchString = paramsQuery.toLowerCase();
-      roomsResult =
-        searchString !== ""
-          ? rooms.filter(
-              ({ divisionName, collegeName, building, code }) =>
-                code.toLowerCase().includes(searchString) ||
-                collegeName?.toLowerCase().includes(searchString) ||
-                divisionName?.toLowerCase().includes(searchString) ||
-                building?.name.toLowerCase().includes(searchString),
-            )
-          : [];
+      // searchInput = paramsQuery;
+      // const searchString = paramsQuery.toLowerCase();
+      // roomsResult =
+      //   searchString !== ""
+      //     ? rooms.filter(
+      //         ({ divisionName, collegeName, building, code }) =>
+      //           code.toLowerCase().includes(searchString) ||
+      //           collegeName?.toLowerCase().includes(searchString) ||
+      //           divisionName?.toLowerCase().includes(searchString) ||
+      //           building?.name.toLowerCase().includes(searchString),
+      //       )
+      //     : [];
     }
 
-    window.addEventListener("keydown", windowKeyDown);
-    return () => {
-      window.removeEventListener("keydown", windowKeyDown);
-    };
+    // window.addEventListener("keydown", windowKeyDown);
+    // return () => {
+    //   window.removeEventListener("keydown", windowKeyDown);
+    // };
   });
 
-  const debounceSearch = debounce((inputValue: string) => {
-    typing = false;
+  // const debounceSearch = debounce((inputValue: string) => {
+  //   typing = false;
 
-    // If the user starts typing a new search while looking at a modal,
-    // close the modal and go back to search view
-    if (inputValue !== "" && modalStore.open) {
-      modalStore.closeModal();
-    }
+  //   if (inputValue !== "" && modalStore.open) {
+  //     modalStore.closeModal();
+  //   }
 
-    roomsResult = findRooms(rooms, inputValue);
-    paginateOffset = 0;
-    const url = new URL(window.location.href);
+  //   // roomsResult = findRooms(rooms, inputValue);
+  //   // paginateOffset = 0;
+  //   const url = new URL(window.location.href);
 
-    if (inputValue === "") url.searchParams.delete("s");
-    else url.searchParams.set("s", inputValue);
+  //   if (inputValue === "") url.searchParams.delete("s");
+  //   else url.searchParams.set("s", inputValue);
 
-    window.history.replaceState({}, "", url);
-  }, 500);
+  //   window.history.replaceState({}, "", url);
+  // }, 500);
 
-  $effect(() => {
-    debounceSearch(searchInput);
-  });
+  // $effect(() => {
+  //   debounceSearch(searchInput);
+  // });
 
-  function windowKeyDown(ev: KeyboardEvent) {
-    if (
-      ev.key === "k" &&
-      ev.ctrlKey &&
-      searchElement != null &&
-      searchElement !== document.activeElement
-    ) {
-      ev.preventDefault();
-      searchElement.focus();
-    }
-  }
-
-  function findRooms(
-    rooms: RoomData[],
-    searchTerm: string,
-    // { type, filter }: Pick<IFilterStore, "filter" | "type">,
-  ) {
-    searchTerm = searchTerm.toLowerCase().trim();
-
-    // if (filter !== null)
-    //   switch (type) {
-    //     case "building":
-    //       rooms = rooms.filter((room) => room.building?.name.includes(filter));
-    //       break;
-    //     case "college":
-    //       rooms = rooms.filter((room) => room.collegeName?.includes(filter));
-    //       break;
-    //     case "division":
-    //       rooms = rooms.filter((room) => room.divisionName?.includes(filter));
-    //       break;
-    //   }
-
-    return searchInput !== ""
-      ? rooms.filter(
-          ({ divisionName, collegeName, building, code }) =>
-            code.toLowerCase().includes(searchTerm) ||
-            collegeName?.toLowerCase().includes(searchTerm) ||
-            divisionName?.toLowerCase().includes(searchTerm) ||
-            building?.name.toLowerCase().includes(searchTerm) ||
-            classesMap
-              .get(code)
-              ?.some((c) => c.courseCode.toLowerCase().includes(searchTerm)),
-        )
-      : rooms;
-  }
-
-  function handleInput() {
-    typing = true;
-    if (modalStore.open) {
-      modalStore.closeModal();
-    }
-  }
-
-  // function scrollToTop() {
-  //   const panel = document.querySelector(".panel-content");
-  //   if (panel) {
-  //     panel.scrollTo({
-  //       top: 0,
-  //       behavior: "smooth",
-  //     });
+  // function windowKeyDown(ev: KeyboardEvent) {
+  //   if (
+  //     ev.key === "k" &&
+  //     ev.ctrlKey &&
+  //     searchElement != null &&
+  //     searchElement !== document.activeElement
+  //   ) {
+  //     ev.preventDefault();
+  //     searchElement.focus();
   //   }
   // }
 
-  function closeSearchContext() {
-    modalStore.closeModal();
-    searchInput = "";
-  }
+  // function findRooms(rooms: RoomData[], searchTerm: string) {
+  //   searchTerm = searchTerm.toLowerCase().trim();
+
+  //   return searchInput !== ""
+  //     ? rooms.filter(
+  //         ({ divisionName, collegeName, building, code }) =>
+  //           code.toLowerCase().includes(searchTerm) ||
+  //           collegeName?.toLowerCase().includes(searchTerm) ||
+  //           divisionName?.toLowerCase().includes(searchTerm) ||
+  //           building?.name.toLowerCase().includes(searchTerm) ||
+  //           classesMap
+  //             .get(code)
+  //             ?.some((c) => c.courseCode.toLowerCase().includes(searchTerm)),
+  //       )
+  //     : rooms;
+  // }
 </script>
 
 <div class="side-panel-wrapper">
-  <div class="search-filter-row">
-    <div class="search-container">
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="20"
-        height="20"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        stroke-width="2"
-        stroke-linecap="round"
-        stroke-linejoin="round"
-        class="search-icon"
-        ><circle cx="11" cy="11" r="8" /><line
-          x1="21"
-          y1="21"
-          x2="16.65"
-          y2="16.65"
-        /></svg
-      >
-      <input
-        type="text"
-        id="search"
-        bind:this={searchElement}
-        bind:value={searchInput}
-        class={typing ? "typing" : ""}
-        oninput={handleInput}
-        placeholder={"Search room code, building, division..."}
-      />
-      {#if typing}
-        <div class="loading-icon">
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 200 200"
-            width="20"
-            height="20"
-          >
-            <circle stroke-width="17" r="15" cx="40" cy="65"
-              ><animate
-                attributeName="cy"
-                calcMode="spline"
-                dur="0.8"
-                values="65;135;65;"
-                keySplines=".5 0 .5 1;.5 0 .5 1"
-                repeatCount="indefinite"
-                begin="-.4"
-              ></animate></circle
-            >
-            <circle stroke-width="17" r="15" cx="100" cy="65"
-              ><animate
-                attributeName="cy"
-                calcMode="spline"
-                dur="0.8"
-                values="65;135;65;"
-                keySplines=".5 0 .5 1;.5 0 .5 1"
-                repeatCount="indefinite"
-                begin="-.2"
-              ></animate></circle
-            >
-            <circle stroke-width="17" r="15" cx="160" cy="65"
-              ><animate
-                attributeName="cy"
-                calcMode="spline"
-                dur="0.8"
-                values="65;135;65;"
-                keySplines=".5 0 .5 1;.5 0 .5 1"
-                repeatCount="indefinite"
-                begin="0"
-              ></animate></circle
-            >
-          </svg>
-        </div>
-      {/if}
-    </div>
-
-    <div class="search-buttons">
-      {#if searchInput !== ""}
-        <button
-          onclick={closeSearchContext}
-          type="button"
-          class="clear-btn"
-          aria-label="Clear Search"
-        >
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            width="20"
-            height="20"
-            viewBox="0 0 24 24"
-            fill="none"
-            stroke="currentColor"
-            stroke-width="2"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            ><line x1="18" y1="6" x2="6" y2="18"></line><line
-              x1="6"
-              y1="6"
-              x2="18"
-              y2="18"
-            ></line></svg
-          >
-        </button>
-      {/if}
-    </div>
-  </div>
+  <Search />
 </div>
 
 <style>
@@ -278,80 +123,6 @@
       flex-direction: column;
       justify-content: space-between; /* Space between search top and panel bottom */
     }
-  }
-
-  .search-filter-row {
-    background-color: white;
-    border-radius: 0.8125rem;
-    box-shadow: 0rem 2px 0.25rem 0rem rgba(0, 0, 0, 0.25);
-    display: flex;
-    align-items: center;
-    padding: 0.75rem 1rem;
-    gap: 0.5rem;
-    flex-shrink: 0;
-  }
-
-  @media screen and (max-width: 48rem) {
-    .search-filter-row {
-      pointer-events: auto;
-      border-radius: 2rem; /* Pill shape for mobile */
-    }
-  }
-
-  .search-container {
-    position: relative;
-    flex: 1;
-    display: flex;
-    align-items: center;
-    min-width: 0;
-  }
-
-  .search-icon {
-    position: absolute;
-    left: 0;
-    color: black;
-  }
-
-  input[type="text"] {
-    width: 100%;
-    border: none;
-    outline: none;
-    padding-left: 1.75rem;
-    font-size: 0.875rem;
-    color: black;
-    background: transparent;
-    text-overflow: ellipsis;
-  }
-
-  input[type="text"]::placeholder {
-    color: #6b6b6b;
-  }
-
-  .search-buttons {
-    display: flex;
-    gap: 0.5rem;
-    align-items: center;
-  }
-
-  .clear-btn {
-    all: unset;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: black;
-  }
-
-  .loading-icon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-left: 0.5rem;
-  }
-
-  .loading-icon svg {
-    fill: #7b2d26;
-    stroke: #7b2d26;
   }
 
   /* .panel-content {
